+++ relay-server/src/utils/mod.rs
@@ -20,7 +20,6 @@ mod memory;
 #[cfg(feature = "processing")]
 mod native;
 mod serde;
-#[cfg(feature = "processing")]
 mod unreal;
 
 pub use self::api::*;
@@ -32,7 +31,6 @@ pub use self::multipart::*;
 #[cfg(feature = "processing")]
 pub use self::native::*;
 pub use self::param_parser::*;
-#[cfg(feature = "processing")]
 pub use self::pick::*;
 pub use self::rate_limits::*;
 pub use self::retry::*;
@@ -44,5 +42,4 @@ pub use self::split_off::*;
 pub use self::statsd::*;
 pub use self::stream::*;
 pub use self::thread_pool::*;
-#[cfg(feature = "processing")]
 pub use self::unreal::*;